backend body 1.2.1

+++ src/auctions/models.rs
@@ -1,8 +1,9 @@
-/// src/auctions/models.rs (or handlers.rs at top)
+// src/auctions/models.rs
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 use time::OffsetDateTime;
 use sqlx::FromRow;
+use bigdecimal::BigDecimal;
 
 #[derive(Debug, Serialize, Deserialize, FromRow)]
 pub struct Auction {
@@ -10,12 +11,12 @@ pub struct Auction {
     pub vehicle_id:            Uuid,
     pub start_time:            OffsetDateTime,
     pub end_time:              OffsetDateTime,
-    pub starting_bid:          f64,             // DOUBLE PRECISION
-    pub current_highest_bid:   Option<f64>,     // DOUBLE PRECISION NULLABLE
-    pub highest_bidder_id:     Option<Uuid>,    // UUID NULLABLE
-    pub status:                String,          // TEXT NOT NULL
-    pub created_at:            OffsetDateTime,  // TIMESTAMPTZ NOT NULL
-    pub updated_at:            OffsetDateTime,  // TIMESTAMPTZ NOT NULL
+    pub starting_bid:          BigDecimal,
+    pub current_highest_bid:   Option<BigDecimal>,
+    pub highest_bidder_id:     Option<Uuid>,
+    pub status:                String,
+    pub created_at:            OffsetDateTime,
+    pub updated_at:            OffsetDateTime,
 }
 
 #[derive(Debug, Serialize, Deserialize, FromRow)]
@@ -23,6 +24,6 @@ pub struct Bid {
     pub id:         Uuid,
     pub auction_id: Uuid,
     pub bidder_id:  Uuid,
-    pub bid_amount: f64,              // DOUBLE PRECISION
-    pub bid_time:   OffsetDateTime,   // TIMESTAMPTZ NOT NULL
+    pub bid_amount: BigDecimal,
+    pub bid_time:   OffsetDateTime,
 }
